requirements.txt
fastapi
streamlit
uvicorn 
wikipedia-api 
requests 
python-dotenv 
google-generativeai
wikipedia
langchain-community